datas/backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==4.2.3
django-cors-headers==4.3.1
django-crontab==0.7.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.1
dnspython
drf-pyotp==0.5.6
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jwcrypto==1.5.6
packaging==24.0
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pytz==2024.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
zope.interface==6.3
